move csv parser to util
